+++ src/Views/steps/license.php
@@ -14,7 +14,10 @@
         <?php include \Installer\Core\Utils::getBasePath('src/Views/partials/progressbar.php'); ?>
 
         <div class="license-content" style="height: 300px; overflow-y: scroll; border: 1px solid #dee2e6; padding: 15px; background-color: #f8f9fa;">
-            <?php include \Installer\Core\Utils::getBasePath('LICENSE'); ?>
+            <h5>MIT License</h5>
+            <p>Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:</p>
+            <p>The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.</p>
+            <p>THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.</p>
         </div>
 
         <form action="index.php?step=license" method="POST" class="mt-4">